refactor: attackResponse
